Update software-build-development.yml
Updating github action versions.
Adding setting include-hidden-file to true. 
This option was "fixed" to not include files by default and it broke our zip deploy because there is a hidden .azurefunctions folder that is required for Azure Functions to work.
Without it the function is not loaded and you get a "jobs not found" error in the host.